Redid the Maze.
Now it is a spiraling stair case where we can project the movies onto the walls as the player ascends his/her way up.

At the end is a projector, and the final trigger should activate the ending, which will play out on a movie screen.  Possibly reconstruct the movie theatre and the player is looking at the screen from above in the projector room.

WIP.